reorg to start measures/trials
